headroom fixes - though still janky
I think we maybe need to ditch headroom as it's not un-freezing the nav 
at the right time. Seems to be pretty sluggish, and the class selectors 
required to get the desired behavior are bloated.

This problem has been 
solved countless times on other sites so let's try not to reinvent the 
wheel here and just find a solution that's fit for purpose.